refactor(plan): unify StudyPlan type into lib/types/studyPlan.ts (#189 prep)
Resolves the duplicate StudyPlan/MonthlyGoal definitions previously kept

in lib/api.ts (no id) and components/features/dashboard/GoalSettingWizard.tsx

(id required). Both modules now re-export from the single canonical type.

Per design doc 21_StudyPlanEditUI.md ｧ4.
